Move LevelsEditor to components
