+++ src/components/LevelsEditor/LevelsEditor.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { toDataFrame } from '@grafana/data';
 import { Select } from '@grafana/ui';
 import { act, fireEvent, render, screen, within } from '@testing-library/react';
-import { TestIds } from '../../../../constants';
+import { TestIds } from '../../constants';
 import { LevelsEditor } from './LevelsEditor';
 
 /**
