+++ src/components/LevelsEditor/styles.ts
